Nothing (but read the commit notes)
This is a commit so I can say that there is a error with the talons (shocker). The robot max tested on ran the auto just fine with the code, this is the same code I ran on our robot and the encoders read fine our calulations came out right, but insted of going 120 inches it went 98 on the left and the 108 right. More work on the talons is required.